Partial implementation of vocab in binary format.
git-svn-id: file:///dev/shm/somefilter.svn@440 e102df66-1e2e-11dd-9b44-c24451a4db5e

+++ lm/binary_format.cc
@@ -144,7 +144,7 @@ uint8_t *SetupZeroed(const Config &config, ModelType model_type, const std::vect
 }
 
 void SizeCheck(std::size_t expected, const uint8_t *start, const Backing &backing) {
-  if (expected < backing.memory.end() - start) {
+  if (expected < static_cast<std::size_t>(backing.memory.end() - start)) {
     UTIL_THROW(FormatLoadException, "Binary file should have size >=" << (expected + start - backing.memory.begin()) << " based on the counts and configuration.");
   }
 }
